+++ lua/plugins/snippets.lua
@@ -1,0 +1,22 @@
+-- ------------------------------------------
+-- Snippets plugins
+--
+-- Override or add snippet related plugins & configuration here
+-- ------------------------------------------
+
+-- if true then return {} end -- WARN: REMOVE THIS LINE TO ACTIVATE THIS FILE
+
+---@type LazySpec
+return {
+  --LuaSnip with json format snippets in `snippets/` directory
+  {
+    "L3MON4D3/LuaSnip",
+    config = function(plugin, opts)
+      require "astronvim.plugins.configs.luasnip"(plugin, opts) -- include the default astronvim config that calls the setup call
+      -- add more custom luasnip configuration such as filetype extend or custom snippets
+      require("luasnip.loaders.from_vscode").lazy_load { paths = { "./snippets" } } -- include JSON style snippets
+      local luasnip = require "luasnip"
+      luasnip.filetype_extend("javascript", { "javascriptreact" })
+    end,
+  },
+}