+++ CHANGELOG.md
@@ -20,5 +20,6 @@
 - mapping: set VM_leader for visual-multi plugin for multiple cursors
 - community: noice UI replacement for messages, popups & command line
 - mapping: add `SPC W` save as key binding prompting for file name
+- snippets: luasnip plugin and json format snippets
 
 ## Changed